# PERF: use non-copying path for Groupby.skew