--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -973,9 +973,12 @@ def group_skew(
                 elif M2[i, j] == 0:
                     out[i, j] = 0
                 else:
-                    out[i, j] = <float64_t>(
-                        (ct * (ct - 1) ** 0.5 / (ct - 2)) * (M3[i, j] / M2[i, j]**1.5)
-                    )
+                    term1 = (ct - 1) ** 0.5
+                    term1 = ct * term1
+                    term1 = term1 / (ct - 2)
+                    term1 = term1 * M3[i, j]
+                    term1 = term1 / M2[i, j]**1.5
+                    out[i, j] = term1
 
 
 @cython.wraparound(False)