--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -977,7 +977,7 @@ def group_skew(
                     term1 = ct * term1
                     term1 = term1 / (ct - 2)
                     term1 = term1 * M3[i, j]
-                    term1 = term1 / M2[i, j]**1.5
+                    term1 = term1 / <float64_t>(M2[i, j]**1.5)
                     out[i, j] = term1