--- a/pandas/_libs/groupby.pyx
+++ b/pandas/_libs/groupby.pyx
@@ -894,6 +894,7 @@ def group_var(
 @cython.wraparound(False)
 @cython.boundscheck(False)
 @cython.cdivision(True)
+@cython.cpow
 def group_skew(
     float64_t[:, ::1] out,
     int64_t[::1] counts,
@@ -911,7 +912,7 @@ def group_skew(
         float64_t[:, ::1] M1, M2, M3
         float64_t delta, delta_n, term1, val
         int64_t n1, n
-        float64_t ct, m2_15, root_n
+        float64_t ct
 
     if len_values != len_labels:
         raise ValueError("len(index) != len(labels)")
@@ -972,10 +973,9 @@ def group_skew(
                 elif M2[i, j] == 0:
                     out[i, j] = 0
                 else:
-                    m2_15 = <float64_t>(M2[i, j] ** 1.5)
-                    root_n = <float64_t>((ct - 1) ** 0.5)
                     out[i, j] = (
-                        (ct * root_n / (ct - 2)) * (M3[i, j] / m2_15)
+                        (ct * (ct - 1) ** 0.5 / (ct - 2))
+                        * (M3[i, j] / M2[i, j] ** 1.5)
                     )