It wouldn't surprise me if there are contiguity-related lookup-optimizations that could be done in group_skew.

```
import pandas as pd
import numpy as np

nrows = 10**5
ngroups = 5
ncols = 4
nan_frac = 0.05

arr = np.random.randn(nrows, ncols)
arr[np.random.random(nrows) < nan_frac] = np.nan

df = pd.DataFrame(arr)
grps = np.random.randint(0, ngroups, size=nrows)
gb = df.groupby(grps)

%timeit result = gb.skew()
7.75 ms ± 25.7 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <-main
1.78 ms ± 115 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  # <- PR

[same with ngroups=50]
%timeit result = gb.skew()
23.9 ms ± 101 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  # <- main
1.81 ms ± 51 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)  # <- PR
```