@WillAyd 2023-03-23
It looks like Cython 3.0 has a new cpow directive that I think is responsible for this behavior. Might be ok to set that to True