@jbrockmendel 2023-03-29
IIUC the cpow thing has to wait until we use cython3. anything else here?